Split out 'pure' introspection lib
