+++ dynmsg/src/string_utils.cpp
@@ -12,18 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DYNMSG_DEMO_INCLUDE_DYNMSG_DEMO_STRING_UTILS_HPP_
-#define DYNMSG_DEMO_INCLUDE_DYNMSG_DEMO_STRING_UTILS_HPP_
+#include "dynmsg/string_utils.hpp"
 
-#include <string>
+#include <codecvt>
+#include <locale>
 
-extern "C"
-{
+std::u16string string_to_u16string(const std::string& input) {
+  std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> converter;
+  return converter.from_bytes(input);
+}
 
-// Convert a std::string (8-bit characters) to a std::u16string (16-bit characters)
-std::u16string string_to_u16string(const std::string& input);
-// Convert a std::u16string (16-bit characters) to a std::string (8-bit characters)
-std::string u16string_to_string(const std::u16string& input);
 
-};  // extern "C"
-#endif  // DYNMSG_DEMO_INCLUDE_DYNMSG_DEMO_STRING_UTILS_HPP_
+std::string u16string_to_string(const std::u16string& input) {
+  std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> converter;
+  return converter.to_bytes(input);
+}
