+++ dynmsg_demo/test/msg_parser_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "dynmsg_demo/msg_parser.hpp"
+#include "dynmsg/msg_parser.hpp"
 #include "dynmsg_demo/typesupport_utils.hpp"
 
 #include <dynmsg_msgs/msg/wide_string.h>
